build: end-to-end build verified on macOS arm64
First full local build: ORB-SLAM3 submodule (with all 9 patches) + Pangolin
(built into conda env) + our wrapper (libspatial_publisher + mono_tum).
mono_tum binary launches cleanly, dyld resolves every dependency, prints
usage and exits.

Build chain (from a fresh `conda env create -f environment.yml`):
  1. scripts/build-pangolin.sh    Pangolin 0.9.5 -> $CONDA_PREFIX
  2. scripts/build-orb-slam3.sh   patches + DBoW2 + g2o + Sophus + libORB_SLAM3
  3. scripts/build.sh             our CMake -> mono_tum (104K)
Total wall time on M-series Mac: ~6 min (Pangolin) + ~4 min (ORB-SLAM3) +
~30 s (wrapper).

Patches added during this build:
  07 — orbslam-cxx17           Bump CMake C++11 -> C++17 (Pangolin headers
                               require std::enable_if_t).
  08 — loopclosing-cxx17-bool  LoopClosing::mnFullBAIdx was declared bool
                               but used with `++` and assigned to int; the
                               `bool++` is removed in C++17. Change type to
                               int (matches all other usages).
  09 — map-friend-access       Map::RemapIds() (added by patch 04) reads
                               KeyFrame/MapPoint protected backup-id fields,
                               so KeyFrame and MapPoint declare `friend
                               class Map`.

Other adjustments:
  - environment.yml pins cmake=3.31.* (cmake 4.x rejects the submodule's
    `cmake_minimum_required(VERSION 2.8)` directives).
  - scripts/build-orb-slam3.sh passes `-DBUILD_TESTS=OFF` to Sophus (its
    test code uses deprecated copy-assignment patterns now hard errors
    under -Werror,-Wdeprecated-copy on modern clang).
  - CMakeLists.txt finds OpenGL explicitly and adds Pangolin include dirs +
    OpenGL::GL link to spatial_publisher (Pangolin transitive header chain
    pulls in <GL/gl.h>, which on macOS lives inside OpenGL.framework and
    needs the conda env's cmake to be told the SDK sysroot).
  - src/socket_publisher.cc include path corrected to lowercase
    "socket_publisher.h".
  - src/main_tum.cc PublishKeyframePose argument is Sophus::SE3f's matrix()
    directly (Eigen::Matrix4f), not a cast<double>() copy.
  - thirdparty/Pangolin added as shallow submodule (auto-cloned by
    scripts/build-pangolin.sh on first run).

Status: precision parity test against the fork's mono_tum is the next
milestone (run both on TUM desk dataset, diff perframe_points.bin).

+++ src/main_tum.cc
@@ -261,7 +261,7 @@ int main(int argc, char** argv)
                     // frame.
                     const bool loopClosed = SLAM.MapChanged();
                     publisher->PublishKeyframePose(
-                        Tcw.matrix().cast<double>(), tframe, loopClosed, state);
+                        Tcw.matrix(), tframe, loopClosed, state);
                 }
             }
         }
